refactor: (#3) s3 package 구조 개선
