+++ src/main/java/com/stac/daijin/domain/card/service/SaveCardService.java
@@ -7,8 +7,8 @@ import com.stac.daijin.domain.user.domain.type.Role;
 import com.stac.daijin.domain.user.domain.User;
 import com.stac.daijin.domain.user.exception.UserPermissionException;
 import com.stac.daijin.domain.user.facade.UserFacade;
-import com.stac.daijin.thirdparty.s3.UploadS3Service;
-import com.stac.daijin.thirdparty.s3.enums.Directory;
+import com.stac.daijin.thirdparty.s3.service.UploadS3Service;
+import com.stac.daijin.thirdparty.s3.type.Directory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
