+++ src/main/java/com/stac/daijin/domain/card/service/UpdateCardService.java
@@ -6,8 +6,8 @@ import com.stac.daijin.domain.card.presentation.dto.request.UpdateCardRequest;
 import com.stac.daijin.domain.user.domain.User;
 import com.stac.daijin.domain.user.exception.IsNotWriterException;
 import com.stac.daijin.domain.user.facade.UserFacade;
-import com.stac.daijin.thirdparty.s3.UploadS3Service;
-import com.stac.daijin.thirdparty.s3.enums.Directory;
+import com.stac.daijin.thirdparty.s3.service.UploadS3Service;
+import com.stac.daijin.thirdparty.s3.type.Directory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
